Refactor QueryType to API
