+++ decidim-core/lib/decidim/core/engine.rb
@@ -365,8 +365,6 @@ module Decidim
       end
 
       initializer "decidim_core.graphql_api" do
-        Decidim::Api::QueryType.include Decidim::QueryExtensions
-
         Decidim::Api.add_orphan_type Decidim::Core::UserType
       end
